--- client/src/components/header/index.js
@@ -1,9 +1,0 @@
-import LoginHeader from "./LoginHeader";
-import LogoutHeader from "./LogoutHeader";
-function Header() {
-  const isLogin = localStorage.getItem("token");
-
-  return <>{isLogin ? <LoginHeader /> : <LogoutHeader />}</>;
-}
-
-export default Header;
